// movie-sphere-ui/src/app/services/models/page-response-movie.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Movie } from '../models/movie';
export interface PageResponseMovie {
  content?: Array<Movie>;
  first?: boolean;
  last?: boolean;
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}